+++ processor/kernel/src/processor/message.rs
@@ -3,3 +3,9 @@ pub enum Content {
     Char(char),
 }
 
+impl Content {
+    pub fn char(character: char) -> Self {
+        Self::Char(character)
+    }
+}
+
